Fix: fix abi stability issue with FResult and ApplicationCommand
The plugin library uses `abi_stable` crate to ensure stable layout of
the types that were used in FResult and ApplicationCommand. However, due
to those structs not using `#[repr(C)]`, their own layout was unstable.
This caused segfaults and memory corruption. For example, a version of
Findex compiled with rustc 1.65 was not compatible with any plugin compiled
with rustc 1.67. From now on, FResult and ApplicationCommand uses `repr(C)` to
ensure a stable layout between different compiler versions.